+++ RapidReactOfficial/src/main/deploy/pathplanner/Double Shot 1.path
@@ -1,1 +1,1 @@
-{"waypoints":[{"anchorPoint":{"x":8.0,"y":2.0},"prevControl":null,"nextControl":{"x":8.145375851291439,"y":1.3964077903710617},"holonomicAngle":90.0,"isReversal":false,"velOverride":null,"isLocked":false},{"anchorPoint":{"x":7.6503241482974085,"y":0.7813241041490006},"prevControl":{"x":7.599138333035153,"y":1.300701812826294},"nextControl":null,"holonomicAngle":90.0,"isReversal":false,"velOverride":null,"isLocked":false}]}
+{"waypoints":[{"anchorPoint":{"x":7.597083570631985,"y":1.8646061950327026},"prevControl":null,"nextControl":{"x":7.583685291882302,"y":1.6040569263755715},"holonomicAngle":90.0,"isReversal":false,"velOverride":null,"isLocked":false},{"anchorPoint":{"x":7.614045359926657,"y":0.7787744872364745},"prevControl":{"x":7.483407491229045,"y":0.7859303438333439},"nextControl":null,"holonomicAngle":87.10955074829154,"isReversal":false,"velOverride":null,"isLocked":false}]}
