+++ RapidReactOfficial/src/main/java/frc/robot/RobotContainer.java
@@ -13,21 +13,21 @@ import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.commands.AimDrivetrainUsingVisionCommand;
-import frc.robot.commands.DriveWithFieldCentricToggleCommand;
-import frc.robot.commands.ExpelAllCommand;
-import frc.robot.commands.FollowTrajectoryCommand;
-import frc.robot.commands.HoundCargo;
-import frc.robot.commands.IntakeCommand;
-import frc.robot.commands.JogIndexRampCommand;
-import frc.robot.commands.JogIndexRampReverseCommand;
-import frc.robot.commands.JogIndexVerticalCommand;
-import frc.robot.commands.JogIndexVerticalReverseCommand;
-import frc.robot.commands.RevShooterAtAutoVelocityCommand;
-import frc.robot.commands.RevShooterAtManualVelocityCommand;
-import frc.robot.commands.RunIndexToShootCommand;
+import frc.robot.commands.AutoAimCommands.AimDrivetrainUsingVisionCommand;
 import frc.robot.commands.AutonomousCommands.DoubleShotTaxiAutonomous;
 import frc.robot.commands.AutonomousCommands.IdealAutonomous;
+import frc.robot.commands.DriveCommands.DriveWithFieldCentricToggleCommand;
+import frc.robot.commands.DriveCommands.FollowTrajectoryCommand;
+import frc.robot.commands.IndexCommands.JogIndexRampCommand;
+import frc.robot.commands.IndexCommands.JogIndexRampReverseCommand;
+import frc.robot.commands.IndexCommands.JogIndexVerticalCommand;
+import frc.robot.commands.IndexCommands.JogIndexVerticalReverseCommand;
+import frc.robot.commands.IndexCommands.RunIndexToShootCommand;
+import frc.robot.commands.IntakeCommands.HoundCargo;
+import frc.robot.commands.IntakeCommands.IntakeCommand;
+import frc.robot.commands.MiscellanousCommands.ExpelAllCommand;
+import frc.robot.commands.ShooterCommands.RevShooterAtAutoVelocityCommand;
+import frc.robot.commands.ShooterCommands.RevShooterAtManualVelocityCommand;
 import frc.robot.subsystems.AutoAimSubsystem;
 import frc.robot.subsystems.ClimberSubsystem;
 import frc.robot.subsystems.DriveSubsystem;
