+++ RapidReactOfficial/src/main/java/frc/robot/commands/AutoAimCommands/PanUsingVisionCommand.java
@@ -2,37 +2,42 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.AutoAimCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.ShooterSubsystem;
+import frc.robot.subsystems.AutoAimSubsystem;
 
-public class RevShooterAtManualVelocityCommand extends CommandBase {
+public class PanUsingVisionCommand extends CommandBase {
 
-  ShooterSubsystem shooterSubsystem;
+  AutoAimSubsystem autoAimSubsystem;
 
-  /** Creates a new ShootAtManuallySelectedVelocityCommand. */
-  public RevShooterAtManualVelocityCommand(ShooterSubsystem shooterSubsystem) {
-    this.shooterSubsystem = shooterSubsystem;
-    addRequirements(shooterSubsystem);
+  public PanUsingVisionCommand(AutoAimSubsystem autoAimSubsystem) {
+    this.autoAimSubsystem = autoAimSubsystem;
+    addRequirements(autoAimSubsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-
+    autoAimSubsystem.enableLimelightLed();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    shooterSubsystem.setFlywheelTargetRpm(shooterSubsystem.getCurrentManualVelocity());
+    autoAimSubsystem.enableLimelightLed();
+    if(autoAimSubsystem.isTargetSeen()) {
+      autoAimSubsystem.setPan(autoAimSubsystem.calculatePanOutput(autoAimSubsystem.getLimelightXOffset()));
+    } else {
+      autoAimSubsystem.stopPan();
+    }
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    shooterSubsystem.stop();
+    autoAimSubsystem.disableLimelightLed();
+    autoAimSubsystem.stopPan();
   }
 
   // Returns true when the command should end.
