+++ RapidReactOfficial/src/main/java/frc/robot/commands/AutonomousCommands/IdealAutonomous.java
@@ -8,9 +8,9 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.PrintCommand;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.commands.AutoAimAutonomousCommand;
-import frc.robot.commands.RunIndexToShootCommand;
-import frc.robot.commands.SetShooterToVelocityCommand;
+import frc.robot.commands.AutoAimCommands.AutoAimAutonomousCommand;
+import frc.robot.commands.IndexCommands.RunIndexToShootCommand;
+import frc.robot.commands.ShooterCommands.SetShooterToVelocityCommand;
 import frc.robot.subsystems.AutoAimSubsystem;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.IndexSubsystem;
