+++ RapidReactOfficial/src/main/java/frc/robot/commands/IndexCommands/JogIndexRampReverseCommand.java
@@ -2,22 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.IndexCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.ConstantsValues;
 import frc.robot.subsystems.IndexSubsystem;
 
-/**
- * Command that runs the vertical part of the index to effectively shoot the ball
- */
-public class RunIndexToShootCommand extends CommandBase {
-
+public class JogIndexRampReverseCommand extends CommandBase {
+  
   IndexSubsystem indexSubsystem;
-  boolean isFinished = false;
 
-  /** Creates a new RunVerticalToShootCommand. */
-  public RunIndexToShootCommand(IndexSubsystem indexSubsystem) {
+  /** Creates a new JogIndexRampReverseCommand. */
+  public JogIndexRampReverseCommand(IndexSubsystem indexSubsystem) {
     this.indexSubsystem = indexSubsystem;
     addRequirements(indexSubsystem);
   }
@@ -29,19 +25,20 @@ public class RunIndexToShootCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    indexSubsystem.setVertical(ConstantsValues.verticalShootSpeed);
-    indexSubsystem.setRamp(ConstantsValues.rampJogSpeed);
-    indexSubsystem.setTransfer(ConstantsValues.transferJogSpeed);
-    isFinished = true;
+    indexSubsystem.setRamp(-ConstantsValues.rampJogSpeed);
+    indexSubsystem.setTransfer(-ConstantsValues.transferJogSpeed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    indexSubsystem.stopRamp();
+    indexSubsystem.stopTransfer();
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return isFinished;
+    return false;
   }
 }
