+++ RapidReactOfficial/src/main/java/frc/robot/commands/IndexCommands/JogIndexVerticalCommand.java
@@ -2,18 +2,18 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.IndexCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.ConstantsValues;
 import frc.robot.subsystems.IndexSubsystem;
 
-public class JogIndexVerticalReverseCommand extends CommandBase {
+public class JogIndexVerticalCommand extends CommandBase {
 
   IndexSubsystem indexSubsystem;
 
-  /** Creates a new JogIndexVerticalReverseCommand. */
-  public JogIndexVerticalReverseCommand(IndexSubsystem indexSubsystem) {
+  /** Creates a new JogIndexVerticalCommand. */
+  public JogIndexVerticalCommand(IndexSubsystem indexSubsystem) {
     this.indexSubsystem = indexSubsystem;
     addRequirements(indexSubsystem);
   }
@@ -25,7 +25,7 @@ public class JogIndexVerticalReverseCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    indexSubsystem.setVertical(-ConstantsValues.verticalJogSpeed);
+    indexSubsystem.setVertical(ConstantsValues.verticalJogSpeed);
   }
 
   // Called once the command ends or is interrupted.
