+++ RapidReactOfficial/src/main/java/frc/robot/commands/IndexCommands/JogIndexVerticalReverseCommand.java
@@ -2,25 +2,19 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.IndexCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.ConstantsValues;
 import frc.robot.subsystems.IndexSubsystem;
-import frc.robot.subsystems.IntakeSubsystem;
 
-public class IntakeCommand extends CommandBase {
+public class JogIndexVerticalReverseCommand extends CommandBase {
 
-  IntakeSubsystem intakeSubsystem;
   IndexSubsystem indexSubsystem;
 
-  boolean isFinished = false;
-
-  /** Creates a new IntakeCommand. */
-  public IntakeCommand(IntakeSubsystem intakeSubsystem, IndexSubsystem indexSubsystem) {
-    this.intakeSubsystem = intakeSubsystem;
+  /** Creates a new JogIndexVerticalReverseCommand. */
+  public JogIndexVerticalReverseCommand(IndexSubsystem indexSubsystem) {
     this.indexSubsystem = indexSubsystem;
-    addRequirements(intakeSubsystem);
     addRequirements(indexSubsystem);
   }
 
@@ -31,20 +25,18 @@ public class IntakeCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    intakeSubsystem.setPower(ConstantsValues.intakePower);
-    indexSubsystem.setRamp(ConstantsValues.rampIntakeSpeed);
-    indexSubsystem.setTransfer(ConstantsValues.transferIntakeSpeed);
-    isFinished = true;
-    
+    indexSubsystem.setVertical(-ConstantsValues.verticalJogSpeed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {}
+  public void end(boolean interrupted) {
+    indexSubsystem.stopVertical();
+  }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return isFinished;
+    return false;
   }
 }
