+++ RapidReactOfficial/src/main/java/frc/robot/commands/IndexCommands/RunIndexToShootCommand.java
@@ -2,18 +2,22 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.IndexCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.ConstantsValues;
 import frc.robot.subsystems.IndexSubsystem;
 
-public class JogIndexVerticalCommand extends CommandBase {
+/**
+ * Command that runs the vertical part of the index to effectively shoot the ball
+ */
+public class RunIndexToShootCommand extends CommandBase {
 
   IndexSubsystem indexSubsystem;
+  boolean isFinished = false;
 
-  /** Creates a new JogIndexVerticalCommand. */
-  public JogIndexVerticalCommand(IndexSubsystem indexSubsystem) {
+  /** Creates a new RunVerticalToShootCommand. */
+  public RunIndexToShootCommand(IndexSubsystem indexSubsystem) {
     this.indexSubsystem = indexSubsystem;
     addRequirements(indexSubsystem);
   }
@@ -25,18 +29,19 @@ public class JogIndexVerticalCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    indexSubsystem.setVertical(ConstantsValues.verticalJogSpeed);
+    indexSubsystem.setVertical(ConstantsValues.verticalShootSpeed);
+    indexSubsystem.setRamp(ConstantsValues.rampJogSpeed);
+    indexSubsystem.setTransfer(ConstantsValues.transferJogSpeed);
+    isFinished = true;
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-    indexSubsystem.stopVertical();
-  }
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return isFinished;
   }
 }
