+++ RapidReactOfficial/src/main/java/frc/robot/commands/IntakeCommands/IntakeCommand.java
@@ -2,58 +2,49 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.IntakeCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.ConstantsValues;
 import frc.robot.subsystems.IndexSubsystem;
 import frc.robot.subsystems.IntakeSubsystem;
-import frc.robot.subsystems.ShooterSubsystem;
 
-public class ExpelAllCommand extends CommandBase {
+public class IntakeCommand extends CommandBase {
 
   IntakeSubsystem intakeSubsystem;
   IndexSubsystem indexSubsystem;
-  ShooterSubsystem shooterSubsystem;
 
-  /** Creates a new ExpelAll. */
-  public ExpelAllCommand(IntakeSubsystem intakeSubsystem, IndexSubsystem indexSubsystem, ShooterSubsystem shooterSubsystem) {
+  boolean isFinished = false;
+
+  /** Creates a new IntakeCommand. */
+  public IntakeCommand(IntakeSubsystem intakeSubsystem, IndexSubsystem indexSubsystem) {
     this.intakeSubsystem = intakeSubsystem;
     this.indexSubsystem = indexSubsystem;
-    this.shooterSubsystem = shooterSubsystem;
     addRequirements(intakeSubsystem);
     addRequirements(indexSubsystem);
-    addRequirements(shooterSubsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    intakeSubsystem.intakeOut();
-    indexSubsystem.setVertical(ConstantsValues.verticalExpelSpeed);
-    indexSubsystem.setRamp(ConstantsValues.rampExpelSpeed);
-    indexSubsystem.setTransfer(ConstantsValues.transferExpelSpeed);
-    shooterSubsystem.setFlywheelVoltage(ConstantsValues.flywheelExpelVolts);
-
+    intakeSubsystem.setPower(ConstantsValues.intakePower);
+    indexSubsystem.setRamp(ConstantsValues.rampIntakeSpeed);
+    indexSubsystem.setTransfer(ConstantsValues.transferIntakeSpeed);
+    isFinished = true;
+    
   }
 
   // Called once the command ends or is interrupted.
   @Override
-  public void end(boolean interrupted) {
-    intakeSubsystem.stop();
-    indexSubsystem.stopAll();
-    shooterSubsystem.stop();
-  }
+  public void end(boolean interrupted) {}
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return false;
+    return isFinished;
   }
 }
