--- RapidReactOfficial/src/main/java/frc/robot/commands/JogIndexRampReverseCommand.java
@@ -1,44 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.Constants.ConstantsValues;
-import frc.robot.subsystems.IndexSubsystem;
-
-public class JogIndexRampReverseCommand extends CommandBase {
-  
-  IndexSubsystem indexSubsystem;
-
-  /** Creates a new JogIndexRampReverseCommand. */
-  public JogIndexRampReverseCommand(IndexSubsystem indexSubsystem) {
-    this.indexSubsystem = indexSubsystem;
-    addRequirements(indexSubsystem);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {}
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-    indexSubsystem.setRamp(-ConstantsValues.rampJogSpeed);
-    indexSubsystem.setTransfer(-ConstantsValues.transferJogSpeed);
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-    indexSubsystem.stopRamp();
-    indexSubsystem.stopTransfer();
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
