+++ RapidReactOfficial/src/main/java/frc/robot/commands/ShooterCommands/RevShooterAtManualVelocityCommand.java
@@ -2,42 +2,37 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.ShooterCommands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.AutoAimSubsystem;
+import frc.robot.subsystems.ShooterSubsystem;
 
-public class PanUsingVisionCommand extends CommandBase {
+public class RevShooterAtManualVelocityCommand extends CommandBase {
 
-  AutoAimSubsystem autoAimSubsystem;
+  ShooterSubsystem shooterSubsystem;
 
-  public PanUsingVisionCommand(AutoAimSubsystem autoAimSubsystem) {
-    this.autoAimSubsystem = autoAimSubsystem;
-    addRequirements(autoAimSubsystem);
+  /** Creates a new ShootAtManuallySelectedVelocityCommand. */
+  public RevShooterAtManualVelocityCommand(ShooterSubsystem shooterSubsystem) {
+    this.shooterSubsystem = shooterSubsystem;
+    addRequirements(shooterSubsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    autoAimSubsystem.enableLimelightLed();
+
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    autoAimSubsystem.enableLimelightLed();
-    if(autoAimSubsystem.isTargetSeen()) {
-      autoAimSubsystem.setPan(autoAimSubsystem.calculatePanOutput(autoAimSubsystem.getLimelightXOffset()));
-    } else {
-      autoAimSubsystem.stopPan();
-    }
+    shooterSubsystem.setFlywheelTargetRpm(shooterSubsystem.getCurrentManualVelocity());
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    autoAimSubsystem.disableLimelightLed();
-    autoAimSubsystem.stopPan();
+    shooterSubsystem.stop();
   }
 
   // Returns true when the command should end.
